fix bugs related to getting all orders and int argument for add product

+++ ShoppingBot/ShoppingBot/Features/Product/AddProduct/AddProductSlashCommand.cs
@@ -18,7 +18,7 @@ namespace ShoppingBot.Features.Product
         [SlashCommand("add-product", "Add a product")]
         [SlashRequirePermissions(DSharpPlus.Permissions.Administrator)]
         public async Task AddProduct(InteractionContext ctx, [Option("name", "Item name")] string name,
-            [Option("price", "Item price")] double price, [Option("description", "Item description")] string description, [Option("image-url", "Url link to the image")] string imageUrl, [Option("quantity", "Item quantity")]int quantity)
+            [Option("price", "Item price")] double price, [Option("description", "Item description")] string description, [Option("image-url", "Url link to the image")] string imageUrl, [Option("quantity", "Item quantity")]double quantity)
         {
             if(quantity == 0)
             {
@@ -26,7 +26,7 @@ namespace ShoppingBot.Features.Product
             }
             else
             {
-                await Create<Entities.Product>(ctx, new AddProductCommand(name, price, description, imageUrl, ctx.Guild.Id.ToString(), quantity));
+                await Create<Entities.Product>(ctx, new AddProductCommand(name, price, description, imageUrl, ctx.Guild.Id.ToString(), (int)quantity));
             }
         }
     }
